+++ package.json
@@ -4,7 +4,7 @@
   "description": "moment js with precision time",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "mocha"
   },
   "repository": {
     "type": "git",
